feat: refatora pedido
