+++ src/dominio/Pedido/PedidoModificacao.java
@@ -6,5 +6,4 @@ public interface PedidoModificacao {
     void adicionarItem(Produto produto, int quantidade, double valorVenda);
     void removerItem(Produto produto);
     void alterarQuantidadeItem(Produto produto, int novaQuantidade);
-    double getValorTotal();
 }
